refactor: centralize frontend logError into shared z2ui5/cc/Logger module
The logError helper was duplicated near-identically across five frontend
modules (Component, Server, DebugTool, App.controller, View1.controller),
with two diverging signatures (some always wrote an `error: undefined`
field, others omitted it when undefined).

Extract a single z2ui5/cc/Logger module and have the four modules with a
clean sap.ui.define dependency list consume it directly. App.controller
keeps a file-scope helper that delegates to the module (it is shared by
~20 control definitions living in separate define blocks) with an inline
fallback. The new module is registered in the http handler preload map
and regenerated into src/01/03 via the app2abap pipeline.
